requirements.txt
# 深度学习框架
torch>=1.10.0
transformers>=4.20.0

# LLM API
dashscope>=1.10.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0

# 机器学习
scikit-learn>=1.0.0

# 配置管理
pyyaml>=5.4.0

# 日志和工具
tqdm>=4.62.0
backoff>=2.1.0

# 可视化 (可选)
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter (可选)
jupyter>=1.0.0
ipykernel>=6.0.0

# 数据增强 (可选)
# jieba>=0.42.0

# 实验跟踪 (可选)
# wandb>=0.12.0
# tensorboard>=2.8.0
